@@ -4,10 +4,11 @@ using System.Linq;
 using System.Threading.Tasks;
 using DotVVM.Framework.Configuration;
 using DotVVM.Framework.Routing;
+using Microsoft.Extensions.DependencyInjection;
 
 namespace DotvvmMvcIntegration
 {
-    public class DotvvmStartup : IDotvvmStartup
+    public class DotvvmStartup : IDotvvmStartup, IDotvvmServiceConfigurator
     {
         public void Configure(DotvvmConfiguration config, string applicationPath)
         {
@@ -20,6 +21,7 @@ namespace DotvvmMvcIntegration
         {
             config.RouteTable.Add("DotvvmSample", "DotvvmSample", "DotVVM/Views/DotvvmSample.dothtml");
 
+            // URLs that are not registered in DotVVM are passed to the next middleware (MVC)
         }
 
         private void ConfigureControls(DotvvmConfiguration config, string applicationPath)
@@ -30,6 +32,11 @@ namespace DotvvmMvcIntegration
         private void ConfigureResources(DotvvmConfiguration config, string applicationPath)
         {
             // register custom resources and adjust paths to the built-in resources
+        }
+
+        public void ConfigureServices(IDotvvmServiceCollection options)
+        {
+            options.AddDefaultTempStorages("Temp");
         }
     }
 }
